动态、修改密码、部分bugg

+++ app/src/main/java/com/example/sig/lianjiang/model/CommentUser.java
@@ -5,11 +5,11 @@ package com.example.sig.lianjiang.model;
  */
 
 public class CommentUser {
-    public long mId; // id
+    public int mId; // id
     public String mName; // 用户名
 
 
-    public CommentUser(long mId, String mName) {
+    public CommentUser(int mId, String mName) {
         this.mId = mId;
         this.mName = mName;
     }
